move nested modules
